requirements.txt
llama-index
sentence-transformers
typer[all]
rich
ctransformers
faiss-cpu